% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qapproach-package.R
\docType{package}
\name{qapproach-package}
\alias{qapproach-package}
\title{A Q-based workflow to support decision-making processes by identifying shared perspectives}
\description{
Tools to prepare Q-sort rankings, identify group perspectives, calculate
consensus priority scores, validate results, and create figures.
}
\references{
Zabala, A. (2014). qmethod: A Package to Explore Human
Perspectives Using Q Methodology. *The R Journal*, 6(2), 163--173.
\doi{10.32614/RJ-2014-032}

Zabala, A., & Pascual, U. (2016). Bootstrapping Q Methodology to Improve
the Understanding of Human Perspectives. *PLOS ONE*, 11(2), e0148087.
\doi{10.1371/journal.pone.0148087}

Geschke, J., Urbach, D., Prescott, G. W., & Fischer, M. (2022). The Q
approach to consensus building: integrating diverse perspectives to guide
decision-making. *ECOEVORXIV*. \doi{10.32942/X2F59S}
}
\seealso{
Useful links:
\itemize{
  \item \doi{10.5281/zenodo.11518485}
  \item Report bugs at \url{https://github.com/JonasGeschke/qapproach/issues}
}

}
\author{
\strong{Maintainer}: Jonas Geschke \email{hallo@qapproach.app}

Authors:
\itemize{
  \item Jonas Geschke \email{hallo@qapproach.app}
}

}
\keyword{internal}
